float?

+++ EloRatings.py
@@ -44,8 +44,8 @@ class Player:
         return (hypothetialRating, hypothetialOther)
 
     def chance(self, other):
-        expected = 1.0 / (1.0 + math.pow(10, (other.rating - self.rating) / 400.0))
-        otherExpected = 1.0 / (1.0 + math.pow(10, (self.rating - other.rating) / 400.0))
+        expected = float(1.0 / (1.0 + math.pow(10, (other.rating - self.rating) / 400.0)))
+        otherExpected = float(1.0 / (1.0 + math.pow(10, (self.rating - other.rating) / 400.0)))
         return (expected, otherExpected)
 
     def __str__(self):
